Improved Heuristic
Improved the heuristic so that it will devalue pieces adjacent to
corners that it does not own. It also fixed a bug where it thought the
top right corner was adjacent to a corner. Added debugging capabilities
to see how many nodes are expanded: generally, alpha-beta is much faster
than minimax.